refactor: 환경 분리
