+++ backend/src/main/java/coffeeshout/global/config/AwsConfig.java
@@ -9,7 +9,7 @@ import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.presigner.S3Presigner;
 
-@Profile({"!local","!test"})
+@Profile({"!local & !test"})
 @Configuration
 public class AwsConfig {
