+++ backend/src/main/java/coffeeshout/global/config/AwsLocalConfig.java
@@ -3,12 +3,11 @@ package coffeeshout.global.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
-import software.amazon.awssdk.auth.credentials.EnvironmentVariableCredentialsProvider;
 import software.amazon.awssdk.regions.Region;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.presigner.S3Presigner;
 
-@Profile({"local", "test"})
+@Profile({"test", "local"})
 @Configuration
 public class AwsLocalConfig {
 
@@ -16,7 +15,6 @@ public class AwsLocalConfig {
     public S3Client s3Client() {
         return S3Client.builder()
                 .region(Region.AP_NORTHEAST_2)
-                .credentialsProvider(EnvironmentVariableCredentialsProvider.create())
                 .build();
     }
