+++ backend/src/main/java/coffeeshout/room/application/LocalStorageService.java
@@ -6,9 +6,9 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 
+@Profile({"local", "test"})
 @Slf4j
 @Service
-@Profile("local")
 public class LocalStorageService implements StorageService {
 
     private final MeterRegistry meterRegistry;
