+++ backend/src/test/java/coffeeshout/global/config/IntegrationTestConfig.java
@@ -1,24 +1,22 @@
-package coffeeshout.config;
+package coffeeshout.global.config;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Profile;
 import org.springframework.scheduling.TaskScheduler;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
 
 @TestConfiguration
 @Profile("test")
-public class ServiceTestConfig {
+public class IntegrationTestConfig {
 
     @Bean(name = "miniGameTaskScheduler")
-    public TaskScheduler testMiniGameTaskScheduler() {
-        return new TestTaskScheduler();
-    }
-
-    @Bean(name = "delayRemovalScheduler")
-    public TaskScheduler testDelayRemovalScheduler() {
+    public TaskScheduler testIntegrationMiniGameTaskScheduler() {
         return new ShutDownTestScheduler();
     }
 
-
+    @Bean(name = "delayRemovalScheduler")
+    public TaskScheduler testIntegrationDelayRemovalScheduler() {
+        return new ShutDownTestScheduler();
+    }
 }
