+++ backend/src/test/java/coffeeshout/global/config/ServiceTestConfig.java
@@ -1,22 +1,21 @@
-package coffeeshout.config;
+package coffeeshout.global.config;
 
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Profile;
 import org.springframework.scheduling.TaskScheduler;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
 
 @TestConfiguration
 @Profile("test")
-public class IntegrationTestConfig {
+public class ServiceTestConfig {
 
     @Bean(name = "miniGameTaskScheduler")
-    public TaskScheduler testIntegrationMiniGameTaskScheduler() {
-        return new ShutDownTestScheduler();
+    public TaskScheduler testMiniGameTaskScheduler() {
+        return new TestTaskScheduler();
     }
 
     @Bean(name = "delayRemovalScheduler")
-    public TaskScheduler testIntegrationDelayRemovalScheduler() {
+    public TaskScheduler testDelayRemovalScheduler() {
         return new ShutDownTestScheduler();
     }
